feat: HeavyEquipment 엔티티 생성
