+++ sensor/src/main/java/com/iroom/sensor/SensorApplication.java
@@ -1,0 +1,12 @@
+package com.iroom.sensor;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SensorApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SensorApplication.class,args);
+    }
+}
