--- sensor/src/main/java/iroom/domain/HeavyEquipmentRepository.java
@@ -1,13 +1,0 @@
-package iroom.domain;
-
-import iroom.domain.*;
-import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
-//<<< PoEAA / Repository
-@RepositoryRestResource(
-    collectionResourceRel = "heavyEquipments",
-    path = "heavyEquipments"
-)
-public interface HeavyEquipmentRepository
-    extends PagingAndSortingRepository<HeavyEquipment, Long> {}
