--- sensor/src/main/java/iroom/domain/Registered.java
@@ -1,28 +1,0 @@
-package iroom.domain;
-
-import iroom.domain.*;
-import iroom.infra.AbstractEvent;
-import java.time.LocalDate;
-import java.util.*;
-import lombok.*;
-
-//<<< DDD / Domain Event
-@Data
-@ToString
-public class Registered extends AbstractEvent {
-
-    private Long id;
-    private String name;
-    private String type;
-    private String location;
-    private Double radius;
-
-    public Registered(HeavyEquipment aggregate) {
-        super(aggregate);
-    }
-
-    public Registered() {
-        super();
-    }
-}
-//>>> DDD / Domain Event
