--- sensor/src/main/java/iroom/infra/PolicyHandler.java
@@ -1,29 +1,0 @@
-package iroom.infra;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import iroom.config.kafka.KafkaProcessor;
-import iroom.domain.*;
-import javax.naming.NameParser;
-import javax.naming.NameParser;
-import javax.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.stream.annotation.StreamListener;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.stereotype.Service;
-
-//<<< Clean Arch / Inbound Adaptor
-@Service
-@Transactional
-public class PolicyHandler {
-
-    @Autowired
-    WorkerHealthRepository workerHealthRepository;
-
-    @Autowired
-    HeavyEquipmentRepository heavyEquipmentRepository;
-
-    @StreamListener(KafkaProcessor.INPUT)
-    public void whatever(@Payload String eventString) {}
-}
-//>>> Clean Arch / Inbound Adaptor
