+++ sensor/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+spring.application.name=sensor
+
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
